Add remove_command
I think Reactors will interact with a bot by registering and 
unregistering commands whose callbacks lie within the 'realm' of the 
Reactor (whether locally or via RPC).

If so, then I better start fleshing out the CommandBot, perhaps even 
promote it to being the actual Bot – and aliasing CommandBot back to it 
if I do that.