Add basic Backend getter protocol
Something that was oddly lacking. Now that Backend itself is a 
typing.Protocol of sorts, and is expected to keep track of the 
underlying platform (even if the underlying protocol itself is 
stateless), requiring channel and user getters makes sense! :D

Yeah, I feel like breaking backwards compatibility is going to be a 
common theme in 1.0.0. Hence the major version!